+++ tests/test_pattern.py
@@ -1207,6 +1207,10 @@ class TestPattern:
 
         assert sv.isclose(sv_test)
 
+    def test_isolated_nodes(self) -> None:
+        pattern = Pattern(input_nodes=[0, 1], cmds=[E((0, 1)), E((0, 1))])
+        assert pattern.isolated_nodes() == {0, 1}
+
 
 def cp(circuit: Circuit, theta: Angle, control: int, target: int) -> None:
     """Controlled rotation gate, decomposed."""  # noqa: D401
@@ -1256,7 +1260,7 @@ class TestMCOps:
             circuit.rx(v, ANGLE_PI / 9)
 
         pattern = circuit.transpile().pattern
-        graph = pattern.graph()
+        graph = pattern.to_opengraph().graph
 
         graph_ref: nx.Graph[int] = nx.Graph()
         graph_ref.add_nodes_from(range(27))